Add new feature: Faster home page user verification
Problem: when someone tries to access the homepage without logging in
, screen flashes with the home page before code that verifies if there
is a valid suer log in registered.

Cause: this is caused by the script being added at the end of the page,
so it runs after the page renders. the script has to be at the end of
the page since it has a part that requires access to HTML elements,
and those elements need to render before the script runs.

Fix: move the part of code responsible for forwarding user to login
page if verified login info wasn't found in localStorage to a new
script located in the head of the page. this way it runs before page
renders.